Add features to the game.
Add the event listeners for the cells of the drawing area.
Mouseover => black
Mouseout => ease-in to white
Manage to fill evenly the drawing area. The cells dimensions are
calc(100% of the container / number of cell (default to 16)).
Style the body, the header, the manins and add a border to the
container to  make similar to an actual ecth a skect toy.