Style everything up. Apply the chosen color palette by declaring the color in css:root. Add the text for the footer.
Want to change the layout by havingo more like rectangle space
for drawing, damn the time I decided that, because I spend the next
2 hours try to fix the ecuation for the cellSide size. Finally I
discover the problem was the relation size, which was 1:3, so changed
for 1:2 and a little adjust the formula worked (and it  looked
much better).
I didn't had that hard time with maths since highschool.

Lastly fix a bug in the number input, which was defaulting to
black in every event.